:book: [RB-127] Axios interceptors
Move axios interceptors to a separate function.
Fix logged user avatar in the UI.
